refactor: pre-pr 리뷰 Warning 일괄 반영
- REST_API: FAQ 섹션 Postgres 기반으로 갱신, Users POST body에서 provider 제거
- UserRustDataSource.createUser: 서버가 쓰지 않는 providerType/providerUid/email 파라미터 제거
- Widget Intents: performAuthenticatedWidgetPost/sendWidgetRequest 중복 80줄을
  WidgetAuthorizedRequest.swift로 공용화 (logContext로 식별)
- PushSender trait: 부분 성공 시 delivered_tokens 필수 채움 계약을 주석으로 명시
- NOTION_FAQ_GUIDE: "구 버전" 표기 + Postgres 017_faq_table.sql 시드 운영 메모 추가
